chore: updates auth flow
